requirements.txt
selenium==3.141.0
Pillow==7.0.0